split read_potential into 3 different files
split into local, global, external. they are too long to make it one
file.
