+++ mjolnir/input/read_interaction.hpp
@@ -14,7 +14,9 @@
 #include <mjolnir/util/throw_exception.hpp>
 #include <mjolnir/util/get_toml_value.hpp>
 #include <mjolnir/util/logger.hpp>
-#include <mjolnir/input/read_potential.hpp>
+#include <mjolnir/input/read_local_potential.hpp>
+#include <mjolnir/input/read_global_potential.hpp>
+#include <mjolnir/input/read_external_potential.hpp>
 #include <mjolnir/input/read_spatial_partition.hpp>
 #include <memory>
